requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
tqdm>=4.65.0
torch>=2.0.1
transformers>=4.30.2
PyQt6>=6.5.1
scipy>=1.10.1
protobuf>=4.23.3
typing-extensions>=4.6.3